Make GLBO.delete() and GLTexture.delete() idempotent

+++ src/org/helioviewer/jhv/opengl/GLTexture.java
@@ -38,6 +38,8 @@ public class GLTexture {
     }
 
     public void delete() {
+        if (texID == -1)
+            return;
         GL.glDeleteTexture(texID);
         pbo.delete();
         texID = -1;
